feat(jugarCarta): Implement thread every time its a player's turn
Previously, the results of a round would be displayed instantly at the
end of it, making it seem like the only player with an actual turn and
decision-making was the human player. This change adds a 2-second-long
delay for the machines when it is their turn, making it seem like they
are deciding on which card to choose, improving user experience.